FixVersorToEuler

+++ utils.h
@@ -17,7 +17,7 @@ private:
 public:
 	Utilitarios();
 	void convertEulerToVersor(float &rx, float &ry, float &rz, double &ax, double &ay, double &az, double &angle);
-	void convertVersorToEuler(float &x, float &y, float &z, float &w, double &rx, double &ry, double &rz);
+    void convertVersorToEuler(double &x, double &y, double &z, double &w, double &rx, double &ry, double &rz);
 	void unirVectorWithAngle(float &rx, float &ry, float&rz, double &vx, double &vy, double &vz, double &newangle);
 	void compareVols(std::string logfilename, std::string newvolume, std::string imagendef, int numTest);
 	void computeHausdorffDistance(std::string logfilename, typename itk::Image<short int, 3>::ConstPointer a, itk::Image<short int,3>::ConstPointer b);
